backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
